Move internal `checkbox-group.primitive`
